+++ docker/main.go
@@ -18,7 +18,6 @@ import (
 	"context"
 	"dagger/docker/internal/dagger"
 	"fmt"
-	"sync"
 )
 
 type Docker struct {
@@ -101,115 +100,35 @@ func (m *Docker) BuildAndPush(
 	// +optional
 	withDirectories []*dagger.Directory,
 ) (string, error) {
-	// Create a WaitGroup to wait for linting, building, and Trivy scan to complete
-	var wg sync.WaitGroup
-	wg.Add(3) // Wait for 3 goroutines (linting, building, and Trivy scan)
-
-	// Channel to collect errors from goroutines
-	errChan := make(chan error, 3) // Buffer for 3 errors
-
-	// Channel to collect linting results
-	lintResultChan := make(chan string, 1)
-
-	// Channel to collect Trivy scan results
-	trivyResultChan := make(chan string, 1)
-
-	// Step 1: Run linting concurrently
-	go func() {
-		defer wg.Done()
-		lintOutput, err := m.Lint(ctx, source, dockerfile, "error") // Use default threshold
-		if err != nil {
-			errChan <- fmt.Errorf("linting failed: %w", err)
-			lintResultChan <- lintOutput // Send linting output even if there's an error
-			return
-		}
-		lintResultChan <- lintOutput // Send linting output
-		errChan <- nil
-	}()
-
-	// Step 2: Run building concurrently
-	var buildErr error
-	go func() {
-		defer wg.Done()
-		// Debug: Log the source directory
-		fmt.Println("Source Directory:", source)
-
-		// Call the Build function
-		builtImage := m.Build(source, dockerfile, withDirectories)
-		if builtImage == nil {
-			buildErr = fmt.Errorf("build failed: builtImage is nil")
-			errChan <- buildErr
-			return
-		}
-
-		// Debug: Log the built image
-		fmt.Println("Built Image:", builtImage)
-
-		// Push the image to the registry
-		_, err := builtImage.Push(ctx, repositoryName, version, withRegistryUsername, withRegistryPassword, registryUrl)
-		if err != nil {
-			buildErr = fmt.Errorf("push failed: %w", err)
-			errChan <- buildErr
-			return
-		}
-		errChan <- nil
-	}()
-
-	// Step 3: Run Trivy scan on the built image
-	go func() {
-		defer wg.Done()
-		// Wait for the build to complete and check for errors
-		if buildErr != nil {
-			errChan <- fmt.Errorf("Trivy scan skipped due to build failure: %w", buildErr)
-			trivyResultChan <- "Trivy scan skipped due to build failure"
-			return
-		}
-
-		// Construct the fully qualified image reference
-		imageRef := fmt.Sprintf("%s/%s:%s", registryUrl, repositoryName, version)
-
-		// Run Trivy scan on the image reference with registry authentication
-		trivyOutput, err := m.TrivyScan(ctx, imageRef, withRegistryUsername, withRegistryPassword)
-		if err != nil {
-			errChan <- fmt.Errorf("Trivy scan failed: %w", err)
-			trivyResultChan <- trivyOutput // Send Trivy output even if there's an error
-			return
-		}
-		trivyResultChan <- trivyOutput // Send Trivy output
-		errChan <- nil
-	}()
-
-	// Wait for all goroutines to complete
-	wg.Wait()
-
-	// Collect linting results
-	lintOutput := <-lintResultChan
-
-	// Collect Trivy scan results
-	trivyOutput := <-trivyResultChan
-
-	// Collect errors from the channel
-	close(errChan)
-	var errs []error
-	for err := range errChan {
-		if err != nil {
-			errs = append(errs, err)
-		}
+	// Step 1: Run linting
+	lintOutput, err := m.Lint(ctx, source, dockerfile, "error") // Use default threshold
+	if err != nil {
+		return "", fmt.Errorf("linting failed: %w", err)
 	}
-
-	// Output linting results
 	fmt.Println("Linting Results:")
 	fmt.Println(lintOutput)
 
-	// Output Trivy scan results
+	// Step 2: Build the image
+	builtImage := m.Build(source, dockerfile, withDirectories)
+	if builtImage == nil {
+		return "", fmt.Errorf("build failed: builtImage is nil")
+	}
+
+	// Step 3: Push the image to the registry
+	imageRef := fmt.Sprintf("%s/%s:%s", registryUrl, repositoryName, version)
+	_, err = builtImage.Push(ctx, repositoryName, version, withRegistryUsername, withRegistryPassword, registryUrl)
+	if err != nil {
+		return "", fmt.Errorf("push failed: %w", err)
+	}
+
+	// Step 4: Run Trivy scan on the built image
+	trivyOutput, err := m.TrivyScan(ctx, imageRef, withRegistryUsername, withRegistryPassword)
+	if err != nil {
+		return "", fmt.Errorf("Trivy scan failed: %w", err)
+	}
 	fmt.Println("Trivy Scan Results:")
 	fmt.Println(trivyOutput)
 
-	// If there are any errors, return them along with the linting and Trivy results
-	if len(errs) > 0 {
-		return fmt.Sprintf("Linting Results:\n%s\nTrivy Scan Results:\n%s", lintOutput, trivyOutput), fmt.Errorf("errors occurred: %v", errs)
-	}
-
 	// Return success along with the linting and Trivy results
-	return fmt.Sprintf("Successfully built and pushed %s/%s:%s\nLinting Results:\n%s\nTrivy Scan Results:\n%s", registryUrl, repositoryName, version, lintOutput, trivyOutput), nil
+	return fmt.Sprintf("Successfully built and pushed %s\nLinting Results:\n%s\nTrivy Scan Results:\n%s", imageRef, lintOutput, trivyOutput), nil
 }
