+++ go/container.go
@@ -1,0 +1,52 @@
+package main
+
+import (
+	"context"
+	"dagger/go/internal/dagger"
+	"fmt"
+	"strings"
+)
+
+func (m *Go) KoBuild(
+	ctx context.Context,
+	src *dagger.Directory,
+	// +optional
+	// +default="GITHUB_TOKEN"
+	tokenName string,
+	token *dagger.Secret,
+	// +optional
+	// +default="ko.local"
+	repo string,
+	// +optional
+	// +default="."
+	buildArg string,
+	// +optional
+	// +default="v0.17.1"
+	koVersion string,
+	// +optional
+	// +default="true"
+	push string,
+) (string, error) {
+	srcDir := "/src"
+
+	ko := m.
+		GetKoContainer(koVersion).
+		WithDirectory(srcDir, src).
+		WithWorkdir(srcDir)
+
+	// DEFINE THE APPLICATION BUILD COMMAND W/ KO
+	output, err := ko.
+		WithEnvVariable("KO_DOCKER_REPO", repo).
+		WithSecretVariable(tokenName, token).
+		WithExec(
+			[]string{"ko", "build", "--push=" + push, buildArg},
+		).
+		Stdout(ctx)
+	if err != nil {
+		return "", fmt.Errorf("error running ko build: %w", err)
+	}
+
+	// Extract the image address from the output
+	imageAddress := strings.TrimSpace(output)
+	return imageAddress, nil
+}
