+++ go/main.go
@@ -274,24 +274,6 @@ func (m *Go) Lint(
 	return dag.GolangciLint().Run(src, golangciLintRunOpts)
 }
 
-// Lint runs the linter on the provided source code
-func (m *Go) ScanTarBallImage(
-	ctx context.Context,
-	file *dagger.File,
-) (*dagger.File, error) {
-	scans := []*dagger.TrivyScan{
-		dag.Trivy().ImageTarball(file),
-	}
-
-	// Grab the report as a file
-	reportFile, err := scans[0].Report("json").Sync(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("error getting report: %w", err)
-	}
-
-	return reportFile, nil
-}
-
 func (m *Go) KoBuildAndScan(
 	ctx context.Context,
 	src *dagger.Directory,
@@ -341,152 +323,6 @@ func (m *Go) KoBuildAndScan(
 	return "No vulnerabilities found.", nil
 }
 
-// SearchVulnerabilities parses the Trivy scan report and filters vulnerabilities by severity
-func (m *Go) SearchVulnerabilities(
-	ctx context.Context,
-	scanOutput string, // The scan output as a string
-	severityFilter string, // Comma-separated list of severities to filter (e.g., "HIGH,CRITICAL")
-) ([]string, error) {
-	// Parse the scan output and filter vulnerabilities by severity
-	var vulnerabilities []string
-
-	// Example: Split the scan output into lines and filter by severity
-	lines := strings.Split(scanOutput, "\n")
-	for _, line := range lines {
-		if strings.Contains(line, severityFilter) {
-			vulnerabilities = append(vulnerabilities, line)
-		}
-	}
-
-	return vulnerabilities, nil
-}
-
-func (m *Go) build(
-	ctx context.Context,
-	src *dagger.Directory,
-	opts GoBuildOpts, // Use the struct for parameters
-) *dagger.Directory {
-	// MOUNT CLONED REPOSITORY INTO `GOLANG` IMAGE
-	golang := m.
-		GetGoLangContainer(opts.GoVersion).
-		WithDirectory("/src", src).
-		WithWorkdir("/src")
-
-	fmt.Println("DIR", src)
-
-	// DEFINE THE APPLICATION BUILD COMMAND
-	path := "build/"
-	buildCmd := []string{
-		"env",
-		"GOOS=" + opts.Os,
-		"GOARCH=" + opts.Arch,
-		"go",
-		"build",
-		"-o",
-		path + "/" + opts.BinName,
-	}
-
-	// Add ldflags if provided
-	if opts.Ldflags != "" {
-		buildCmd = append(buildCmd, "-ldflags", opts.Ldflags)
-	}
-
-	// Add the main Go file to the build command
-	buildCmd = append(buildCmd, opts.GoMainFile)
-
-	// Execute the build command
-	golang = golang.WithExec(buildCmd)
-
-	// GET REFERENCE TO BUILD OUTPUT DIRECTORY IN CONTAINER
-	outputDir := golang.Directory(path)
-
-	return outputDir
-}
-func (m *Go) Binary(
-	ctx context.Context,
-	src *dagger.Directory,
-	// +optional
-	// +default="1.23.6"
-	goVersion string,
-	// +optional
-	// +default="linux"
-	os string,
-	// +optional
-	// +default="amd64"
-	arch string,
-	// +optional
-	// +default="main.go"
-	goMainFile string,
-	// +optional
-	// +default="main"
-	binName string,
-	// +optional
-	ldflags string, // Add ldflags as an optional parameter
-) *dagger.Directory {
-	// Call the core build function with the struct
-	return m.build(ctx, src, GoBuildOpts{
-		GoVersion:  goVersion,
-		Os:         os,
-		Arch:       arch,
-		GoMainFile: goMainFile,
-		BinName:    binName,
-		Ldflags:    ldflags, // Pass ldflags to the build function
-	})
-}
-
-func (m *Go) Build(
-	ctx context.Context,
-	src *dagger.Directory,
-	opts GoBuildOpts, // Use the struct for parameters
-) *dagger.Directory {
-	// Call the core build function
-	return m.build(ctx, src, opts)
-}
-
-func (m *Go) KoBuild(
-	ctx context.Context,
-	src *dagger.Directory,
-	// +optional
-	// +default="GITHUB_TOKEN"
-	tokenName string,
-	token *dagger.Secret,
-	// +optional
-	// +default="ko.local"
-	repo string,
-	// +optional
-	// +default="."
-	buildArg string,
-	// +optional
-	// +default="v0.17.1"
-	koVersion string,
-	// +optional
-	// +default="true"
-	push string,
-) (string, error) {
-	srcDir := "/src"
-
-	ko := m.
-		GetKoContainer(koVersion).
-		WithDirectory(srcDir, src).
-		WithWorkdir(srcDir)
-
-	// DEFINE THE APPLICATION BUILD COMMAND W/ KO
-	output, err := ko.
-		WithEnvVariable("KO_DOCKER_REPO", repo).
-		WithSecretVariable(tokenName, token).
-		WithExec(
-			[]string{"ko", "build", "--push=" + push, buildArg},
-		).
-		Stdout(ctx)
-	if err != nil {
-		return "", fmt.Errorf("error running ko build: %w", err)
-	}
-
-	// Extract the image address from the output
-	imageAddress := strings.TrimSpace(output)
-	return imageAddress, nil
-}
-
 func (m *Go) Test(
 	ctx context.Context,
 	src *dagger.Directory,
@@ -510,59 +346,6 @@ func (m *Go) Test(
 	}
 
 	return output, nil
-}
-
-func (m *Go) TrivyScan(
-	ctx context.Context,
-	src *dagger.Directory,
-	// +optional
-	// +default="HIGH,CRITICAL"
-	severity string, // Severity levels to include (e.g., "HIGH,CRITICAL")
-	// +optional
-	// +default="0.59.1"
-	trivyVersion string,
-) (*dagger.File, error) {
-	// Create a container with Trivy installed
-	container := dag.Container().
-		From("aquasec/trivy:"+trivyVersion). // Use the official Trivy image
-		WithDirectory("/src", src).
-		WithWorkdir("/src")
-
-	// Run Trivy to scan the source folder and write the output to a file
-	reportPath := "/src/trivy-report.txt"
-	container = container.
-		WithExec([]string{"sh", "-c", fmt.Sprintf("trivy fs --severity %s /src > %s || true", severity, reportPath)}) // Ignore exit code
-
-	// Get the Trivy report file
-	reportFile := container.File(reportPath)
-
-	// Return the Trivy report file
-	return reportFile, nil
-}
-
-func (m *Go) SecurityScan(
-	ctx context.Context,
-	src *dagger.Directory,
-	// +optional
-	// +default="2.22.1"
-	secureGoVersion string,
-) (*dagger.File, error) {
-	// Create a container with gosec installed
-	container := dag.Container().
-		From("securego/gosec:"+secureGoVersion). // Use the official gosec image
-		WithDirectory("/src", src).
-		WithWorkdir("/src")
-
-	// Run gosec to scan the source code and write the output to a file
-	reportPath := "/src/security-report.txt"
-	container = container.
-		WithExec([]string{"sh", "-c", "gosec ./... > " + reportPath + " || true"}) // Ignore exit code
-
-	// Get the security report file
-	reportFile := container.File(reportPath)
-
-	// Return the security report file
-	return reportFile, nil
 }
 
 func (m *Go) RunWorkflowContainerStage(
@@ -607,23 +390,4 @@ func (m *Go) RunWorkflowContainerStage(
 
 	// Step 4: Return the image address
 	return fmt.Sprintf("Image built and pushed successfully: %s", imageAddress), nil
-}
-
-func (m *Go) ScanRemoteImage(
-	ctx context.Context,
-	imageAddress string, // Remote image address (e.g., "ko.local/my-image:latest")
-	severityFilter string, // Comma-separated list of severities to filter (e.g., "HIGH,CRITICAL")
-) (string, error) {
-	// Create a container with Trivy installed
-	container := dag.Container().
-		From("aquasec/trivy:latest"). // Use the official Trivy image
-		WithExec([]string{"trivy", "image", "--severity", severityFilter, imageAddress})
-
-	// Capture the scan output
-	output, err := container.Stdout(ctx)
-	if err != nil {
-		return "", fmt.Errorf("error running Trivy scan: %w", err)
-	}
-
-	return output, nil
 }
